+++ src/payment/mod.rs
@@ -17,7 +17,7 @@ mod unified_qr;
 
 pub use bolt11::Bolt11Payment;
 pub use bolt12::Bolt12Payment;
-pub use onchain::OnchainPayment;
+pub use onchain::{AddressInfo, KeychainKind, OnchainPayment};
 pub use spontaneous::SpontaneousPayment;
 pub use store::{
 	ConfirmationStatus, LSPFeeLimits, PaymentDetails, PaymentDirection, PaymentKind, PaymentStatus,